pyproject.toml
[project]
name = "mac-capture"
version = "0.1.0"
description = "Minimal macOS ScreenCaptureKit recorder in Python"
requires-python = ">=3.10"
dependencies = [
    "pyobjc",
    "pyobjc-framework-AVFoundation",
    "pyobjc-framework-AudioToolbox",
    "pyobjc-framework-CoreMedia",
    "pyobjc-framework-Quartz",
    "pyobjc-framework-ScreenCaptureKit",
]

[tool.uv]
package = false
